Improve anyOf/oneOf schema docs, parsing, and tests
